Remove analytics API. Fix output diff queries.
