+++ documentation/api/api-explorer.yml
@@ -160,6 +160,57 @@ paths:
           $ref: "#/components/responses/NoResults"
         "500":
           $ref: "#/components/responses/InternalError"
+  /api/analytics/v2/ledger/richest-addresses:
+    get:
+      tags:
+        - ledger
+      summary: Returns the top richest addresses.
+      description: >-
+        Returns analytics about the top richest addresses at the ledger state specified by the provided index
+      parameters:
+        - $ref: "#/components/parameters/ledgerIndex"
+        - $ref: "#/components/parameters/top"
+      responses:
+        "200":
+          description: Successful operation.
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/RichestAddressesResponse"
+              examples:
+                default:
+                  $ref: "#/components/examples/richest-addresses-example"
+        "400":
+          $ref: "#/components/responses/BadRequest"
+        "404":
+          $ref: "#/components/responses/NoResults"
+        "500":
+          $ref: "#/components/responses/InternalError"
+  /api/analytics/v2/ledger/token-distribution:
+    get:
+      tags:
+        - ledger
+      summary: Returns the current token distribution.
+      description: >-
+        Returns analytics about the distribution of IOTA tokens at the ledger state specified by the provided index
+      parameters:
+        - $ref: "#/components/parameters/ledgerIndex"
+      responses:
+        "200":
+          description: Successful operation.
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/TokenDistributionResponse"
+              examples:
+                default:
+                  $ref: "#/components/examples/token-distribution-example"
+        "400":
+          $ref: "#/components/responses/BadRequest"
+        "404":
+          $ref: "#/components/responses/NoResults"
+        "500":
+          $ref: "#/components/responses/InternalError"
 components:
   schemas:
     BalanceResponse:
@@ -285,6 +336,60 @@ components:
           description: The cursor which can be used to retrieve the next logical page of results.
       required:
         - items
+    RichestAddressesResponse:
+      description: Richest addresses statistics.
+      properties:
+        top:
+          type: array
+          description: The top wealthiest addresses.
+          items:
+            type: object
+            properties:
+              address:
+                oneOf:
+                  - $ref: "https://raw.githubusercontent.com/iotaledger/tips/main/tips/TIP-0025/core-rest-api.yaml#/components/schemas/Ed25519Address"
+                  - $ref: "https://raw.githubusercontent.com/iotaledger/tips/main/tips/TIP-0025/core-rest-api.yaml#/components/schemas/AliasAddress"
+                  - $ref: "https://raw.githubusercontent.com/iotaledger/tips/main/tips/TIP-0025/core-rest-api.yaml#/components/schemas/NFTAddress"
+              balance:
+                type: string
+                description: The total balance within this range.
+            required:
+              - address
+              - balance
+      required:
+        - top
+    TokenDistributionResponse:
+      description: Wealth distribution statistics.
+      properties:
+        distribution:
+          type: array
+          description: The distribution of IOTA tokens.
+          items:
+            type: object
+            properties:
+              range:
+                type: object
+                description: The range of balances.
+                properties:
+                  start:
+                    type: number
+                  end:
+                    type: number
+                required:
+                  - start
+                  - end
+              addressCount:
+                type: string
+                description: The number of addresses in this range.
+              totalBalance:
+                type: string
+                description: The total balance within this range.
+            required:
+              - range
+              - addressCount
+              - totalBalance
+      required:
+        - distribution
   responses:
     NoResults:
       description: >-
@@ -371,6 +476,24 @@ components:
         type: string
       example: 1643383242.fa0de75d225cca2799395e5fc340702fc7eac821d2bdd79911126f131ae097a20100.100
       description: The cursor which can be used to retrieve the next logical page of results.
+    ledgerIndex:
+      in: query
+      name: ledgerIndex
+      schema:
+        type: integer
+      example: 200000
+      required: false
+      description: >-
+        The milestone index to be used to determine the ledger state. Defaults to the 
+        application's current ledger index.
+    top:
+      in: query
+      name: top
+      schema:
+        type: number
+      example: 100
+      required: false
+      description: The milestone index to be used to determine the ledger state. Defaults to 200.
   examples:
     balance-example:
       value:
@@ -409,3 +532,34 @@ components:
           - milestoneId: "0xfa0de75d225cca2799395e5fc340702fc7eac821d2bdd79911126f131ae097a2"
             index: 101
         cursor: 102.2
+    richest-addresses-example:
+      value:
+        top:
+          - address:
+              - type: 0
+                pubKeyHash: "0x3845105b59429361a75b3203a6e24e16d19540aad6bd1936449b62f1c4bbe5da"
+            balance: "2779164783277761"
+          - address:
+              - type: 0
+                pubKeyHash: "0xd0d361341fa3bb2f6855039a82ee9ea470c3336eaf34d22767fdfa901ba63e31"
+            balance: "7398600000"
+        ledgerIndex: 1005429
+    token-distribution-example:
+      value:
+        distribution:
+          - range:
+              - start: 100000
+                end: 1000000
+            addressCount: "39"
+            totalBalance: "14612000"
+          - range:
+              - start: 1000000
+                end: 10000000
+            addressCount: "22"
+            totalBalance: "41274500"
+          - range:
+              - start: 100000000
+                end: 1000000000
+            addressCount: "27"
+            totalBalance: "25486528000"
+        ledgerIndex: 1005429
