--- src/bin/inx-chronicle/api/stardust/analytics/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2022 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-mod extractors;
-mod responses;
-mod routes;
-
-pub use self::routes::routes;
