+++ src/bin/inx-chronicle/api/stardust/mod.rs
@@ -3,14 +3,12 @@
 
 use super::router::Router;
 
-pub mod analytics;
 pub mod core;
 pub mod explorer;
 pub mod indexer;
 
 pub fn routes() -> Router {
     Router::new()
-        .nest("/analytics/v2", analytics::routes())
         .nest("/explorer/v2", explorer::routes())
         .nest("/core/v2", core::routes())
         .nest("/indexer/v1", indexer::routes())
